Fix inconsistency in scroll values after changing script
There is an inconsistency in scroll values after changing to a new script. The user has to scroll once to trigger consistency.

This might be because the scroll value of the active CodeEdit after changing script is not updated immediately.

So the present solution is to add a slight delay before the sync.